Add files via upload
In this problem, you will predict the probability of the person or customer defaulting and then go on to classify them as defaulters or non-defaulters. Clean/ transform the data as necessary, perform EDA to analyze the relationship b/w the Status(Target) and the remaining features(Independant). Use a sequential Neural Network model (Feed Forward) to classify the outcome (Defaulter vs Non Defaulter).